🚚 [projects] Move type alias `CreateProject` from `services` to `create` module
